server/model/requirements.txt
tqdm
matplotlib
numpy
pytorch-msssim
mediapipe===0.10.7
opencv-python==4.8.1.78
flask
Werkzeug==2.2.2